Use smallest integer dtype for int props
